Add Video and Project Interfaces to Types
- Introduced new interfaces for Video and Project in types.ts to enhance type safety and structure.
- Updated the projects route to utilize the new VideoGrid and ProjectGrid components, improving modularity and code organization.
- Removed redundant interface definitions from +page.svelte, ensuring adherence to best practices in TypeScript and Svelte development.
