+++ src/lib/types.ts
@@ -14,3 +14,20 @@ export interface EnhancedImage {
 	loading?: 'eager' | 'lazy';
 	decoding?: 'sync' | 'async' | 'auto';
 }
+
+export interface Video {
+	title: string;
+	description: string;
+	topics: string[];
+	url: string;
+	embedId: string;
+	date: string;
+}
+
+export interface Project {
+	title: string;
+	description: string;
+	tech: string[];
+	url: string;
+	status: 'active' | 'stable' | 'experimental';
+}
